+++ CHANGELOG.md
@@ -44,11 +44,13 @@ and this project adheres to
 - @cosmjs/tendermint-rpc: Fix `key` and `value` type in `RpcAbciQueryResponse`
   to also include the `null` option.
 - @cosmjs/tendermint-rpc: Fix decoding events without attributes ([#1198]).
+- @cosmjs/stargate: Return tx response data without ignoring it ([#1241]).
 
 [#1170]: https://github.com/cosmos/cosmjs/issues/1170
 [#1177]: https://github.com/cosmos/cosmjs/issues/1177
 [#1188]: https://github.com/cosmos/cosmjs/pull/1188
 [#1198]: https://github.com/cosmos/cosmjs/pull/1198
+[#1241]: https://github.com/cosmos/cosmjs/pull/1241
 
 ### Changed